fix: separate classes to export as es modules
